Revert "Clean up paths in sample app (#7)"
This reverts commit b81c49ac38327926aabcfdee4cb5df564e046a03.
