+++ basic-app/components/ComponentWithAuth.tsx
