+++ basic-app/pages/api/auth/[...nextauth].ts
