+++ basic-app/pages/app/wallet.tsx
@@ -1,9 +1,9 @@
 import { Box, VStack } from "@chakra-ui/react";
 
-import AppLayout from "components/AppLayout";
-import { AppHeader } from "components/AppHeader";
-import { FlowWalletSetup } from "components/FlowWalletSetup";
-import { ComponentWithAuth } from "components/ComponentWithAuth";
+import AppLayout from "../../components/AppLayout";
+import { AppHeader } from "../../components/AppHeader";
+import { FlowWalletSetup } from "../../components/FlowWalletSetup";
+import { ComponentWithAuth } from "../../components/ComponentWithAuth";
 
 const Drops: ComponentWithAuth = () => (
   <AppLayout>
