+++ basic-app/styles/globals.css
